// src/app/shared/templates/session-handson/models/step-description.ts
/**
 * 專案名稱： course-blog
 * 部門代號： ML8100
 * 檔案說明： 實際演練步驟資料模型
 * @CREATE Sunday, 6th October 2019 5:24:52 pm
 * @author Steve Y Lin
 * @contact Steve_Y_Lin@wistron.com #1342
 * -----------------------------------------------------------------------------
 * @NOTE
 */

/**
 * 實際演練步驟資料模型
 */
export interface StepDescription {
  title: string;
  subtitle?: string;
  description?: string;
}